config/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==22.0.0
httplib2==0.22.0
idna==3.8
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
